+++ config.c
@@ -15,13 +15,25 @@
 #include <unistd.h>
 #include <ctype.h>
 
-
 #define LOG_MODULE "config"
 #define LOG_ENABLE_DBG 0
 #include "log.h"
 #include "key-binding.h"
 
 #define ALEN(v) (sizeof(v) / sizeof((v)[0]))
+
+const struct anchors_map anchors_map[] = {
+    {"top-left", ANCHOR_TOP_LEFT},
+    {"top", ANCHOR_TOP},
+    {"top-right", ANCHOR_TOP_RIGHT},
+    {"left", ANCHOR_LEFT},
+    {"center", ANCHOR_CENTER},
+    {"right", ANCHOR_RIGHT},
+    {"bottom-left", ANCHOR_BOTTOM_LEFT},
+    {"bottom", ANCHOR_BOTTOM},
+    {"bottom-right", ANCHOR_BOTTOM_RIGHT},
+    {NULL, 0},
+};
 
 static const char *const binding_action_map[] = {
     [BIND_ACTION_NONE] = NULL,
@@ -775,6 +787,31 @@ parse_section_main(struct context *ctx)
     else if (strcmp(key, "launch-prefix") == 0)
         return value_to_str(ctx, &conf->launch_prefix);
 
+    else if (strcmp(key, "anchor") == 0) {
+        uint32_t anchor = 0;
+
+        for (size_t i = 0; anchors_map[i].name != NULL; i++) {
+            if (strcmp(value, anchors_map[i].name) == 0) {
+                anchor = anchors_map[i].value;
+                break;
+            }
+        }
+
+        if (anchor == 0) {
+            LOG_CONTEXTUAL_ERR(
+                "invalid anchor \"%s\", "
+                "must be one of: "
+                "\"center\", \"top-left\", \"top\", \"top-right\", "
+                "\"right\", \"bottom-right\", \"bottom\""
+                "\"bottom-left\", \"left\"",
+                value);
+            return false;
+        }
+
+        conf->anchor = anchor;
+        return true;
+    }
+
     else if (strcmp(key, "lines") == 0)
         return value_to_uint32(ctx, 10, &conf->lines);
 
@@ -1414,6 +1451,7 @@ config_load(struct config *conf, const char *conf_path,
             .exit_immediately_if_empty = false,
             .delim = '\n',
         },
+        .anchor = ANCHOR_CENTER,
         .lines = 15,
         .chars = 30,
         .tabs = 8,