+++ main.c
@@ -245,6 +245,7 @@ print_usage(const char *prog_name)
            "  -T,--terminal                  terminal command to use when launching\n"
            "                                 'terminal' programs, e.g. \"xterm -e\".\n"
            "                                 Not used in dmenu mode (not set)\n"
+           "  -a,--anchor                    window anchor (center)\n"
            "  -l,--lines                     number of matches to show\n"
            "  -w,--width                     window width, in characters (margins and\n"
            "                                 borders not included)\n"
@@ -551,6 +552,7 @@ main(int argc, char *const *argv)
         {"no-icons",             no_argument,       0, 'I'},
         {"fields",               required_argument, 0, 'F'},
         {"password",             optional_argument, 0, OPT_PASSWORD},
+        {"anchor",               required_argument, 0, 'a'},
         {"lines",                required_argument, 0, 'l'},
         {"width",                required_argument, 0, 'w'},
         {"tabs",                 required_argument, 0, OPT_TABS},
@@ -606,6 +608,7 @@ main(int argc, char *const *argv)
         bool dpi_aware_set:1;
         bool match_fields_set:1;
         bool icons_disabled_set:1;
+        bool anchor_set:1;
         bool lines_set:1;
         bool chars_set:1;
         bool tabs_set:1;
@@ -655,7 +658,7 @@ main(int argc, char *const *argv)
     }
 
     while (true) {
-        int c = getopt_long(argc, argv, ":o:f:D:IF:il:w:x:y:p:P:b:t:m:s:S:M:B:r:C:T:dRvh", longopts, NULL);
+        int c = getopt_long(argc, argv, ":o:f:D:IF:ia:l:w:x:y:p:P:b:t:m:s:S:M:B:r:C:T:dRvh", longopts, NULL);
         if (c == -1)
             break;
 
@@ -778,6 +781,25 @@ main(int argc, char *const *argv)
 
         case 'T':
             cmdline_overrides.conf.terminal = optarg;
+            break;
+
+        case 'a':
+            uint32_t anchor = 0;
+
+            for (size_t i = 0; anchors_map[i].name != NULL; i++) {
+                if (strcmp(optarg, anchors_map[i].name) == 0) {
+                    anchor = anchors_map[i].value;
+                    break;
+                }
+            }
+
+            if (anchor == 0) {
+                fprintf(stderr, "%s: invalid anchor\n", optarg);
+                return EXIT_FAILURE;
+            }
+
+            cmdline_overrides.conf.anchor = anchor;
+            cmdline_overrides.anchor_set = true;
             break;
 
         case 'l':
@@ -1202,6 +1224,8 @@ main(int argc, char *const *argv)
         conf.match_fields = cmdline_overrides.conf.match_fields;
     if (cmdline_overrides.icons_disabled_set)
         conf.icons_enabled = cmdline_overrides.conf.icons_enabled;
+    if (cmdline_overrides.anchor_set)
+        conf.anchor = cmdline_overrides.conf.anchor;
     if (cmdline_overrides.lines_set)
         conf.lines = cmdline_overrides.conf.lines;
     if (cmdline_overrides.chars_set)