+++ wayland.c
@@ -1298,6 +1298,9 @@ update_size(struct wayland *wayl)
     zwlr_layer_surface_v1_set_size(
         wayl->layer_surface, wayl->width / scale, wayl->height / scale);
 
+    zwlr_layer_surface_v1_set_anchor(
+        wayl->layer_surface, wayl->conf->anchor);
+
     /* Trigger a 'configure' event, after which we'll have the actual width+height */
     wl_surface_commit(wayl->surface);
     wl_display_roundtrip(wayl->display);