+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 ### Added
 
 * Support for the `cursor-shape-v1` Wayland protocol.
+* New option `--anchor` allows setting the window position anchor (i.e. where
+  on the screen the window should generally be positioned), such as `top`,
+  `top-left`, `bottom-right`, `center`, etc. Defaults to `center`
 
 
 ### Changed